Add API.Get Methods for Login Tokens
Add `API.Get.loginTokenWithCred` and `API.Get.loginTokenWithToken`
methods that allow retrieval of a login token that can be used
in place of a user's email and password to authenticate operations.
`loginTokenWithTokens` merely calls `loginTokenWithCred`, but passes
the token as the email and leaves the password blank. Success
and failure are determined by response code and relayed by failure
status in NetworkResponse.

Since the authentication API may change, these methods may too.
However, their behavior and signatures should not unless drastic
changes are made to the authentication API.