Remove `Welcome` App Icon
Before, a `Welcome` App was present on the phone's home screen in
addition to the `Culture Mesh` one. The `Welcome` one led to
`OnboardActivity`, while the `Culture Mesh` one led to `StartActivity`.
To remove the `Welcome` one, `OnboardActivity`'s `<intent-filter>
indicating it was the `MAIN` and `LAUNCHER` activity, was removed from
the manifest.

Resolves #128